Refactor emoji sync to use REST API, remove sharding
Replaces the sharded syncAppEmojis.js and syncEmojiBot.js with a new syncEmojis.js script that uses the Discord REST API directly for emoji management. Updates all related npm scripts to use the new implementation and simplifies the emoji sync process, removing the need for sharding and Discord.js client instantiation.
