+++ discord_bots/package.json
@@ -8,9 +8,9 @@
     "dev:5th:global": "npm run generate-types && node scripts/deployCommands.js 5th dev-global && npm run emoji:sync:5th && nodemon src/shards/index-5th.js",
     "dev:20th:global": "npm run generate-types && node scripts/deployCommands.js 20th dev-global && npm run emoji:sync:20th && nodemon src/shards/index-20th.js",
     "dev:cod:global": "npm run generate-types && node scripts/deployCommands.js cod dev-global && npm run emoji:sync:cod && nodemon src/shards/index-cod.js",
-    "deploy:5th": "node scripts/deployCommands.js 5th prod && node scripts/syncAppEmojis.js 5th",
-    "deploy:20th": "node scripts/deployCommands.js 20th prod && node scripts/syncAppEmojis.js 20th",
-    "deploy:cod": "node scripts/deployCommands.js cod prod && node scripts/syncAppEmojis.js cod",
+    "deploy:5th": "node scripts/deployCommands.js 5th prod && node scripts/syncEmojis.js 5th",
+    "deploy:20th": "node scripts/deployCommands.js 20th prod && node scripts/syncEmojis.js 20th",
+    "deploy:cod": "node scripts/deployCommands.js cod prod && node scripts/syncEmojis.js cod",
     "deploy:all": "npm run deploy:commands:all && npm run emoji:sync:all",
     "deploy:commands:all": "node scripts/deployCommands.js 5th prod && node scripts/deployCommands.js 20th prod && node scripts/deployCommands.js cod prod",
     "clear:5th:dev": "node scripts/deployCommands.js 5th clear-dev",
@@ -21,10 +21,10 @@
     "clear:20th:global": "node scripts/deployCommands.js 20th clear-global",
     "clear:cod:global": "node scripts/deployCommands.js cod clear-global",
     "clear:all:global": "node scripts/deployCommands.js 5th clear-global && node scripts/deployCommands.js 20th clear-global && node scripts/deployCommands.js cod clear-global",
-    "emoji:sync:5th": "node scripts/syncAppEmojis.js 5th",
-    "emoji:sync:20th": "node scripts/syncAppEmojis.js 20th",
-    "emoji:sync:cod": "node scripts/syncAppEmojis.js cod",
-    "emoji:sync:all": "node scripts/syncAllEmojis.js",
+    "emoji:sync:5th": "node scripts/syncEmojis.js 5th",
+    "emoji:sync:20th": "node scripts/syncEmojis.js 20th",
+    "emoji:sync:cod": "node scripts/syncEmojis.js cod",
+    "emoji:sync:all": "node scripts/syncEmojis.js 5th && node scripts/syncEmojis.js 20th && node scripts/syncEmojis.js cod",
     "generate-types": "node scripts/generateEmojiTypes.js",
     "test": "jest",
     "format": "prettier --write \"src/**/*.{js,ts,json}\"",
